backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
PyMySQL
tensorflow
Pillow
numpy
Werkzeug
opencv-python
Flask-Limiter
gunicorn
cryptography
